Development

+++ src/core/adapters/controllers/BookingController.ts
@@ -1,4 +1,4 @@
-import { NextFunction, Request, Response } from "express";
+import { Request, Response } from "express";
 import { inject, injectable } from "inversify";
 import createError from "http-errors";
 
@@ -10,7 +10,7 @@ import {
   offsetLimitSchema,
   updateBookingSchema,
 } from "../../../validations";
-import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
+import logger from "../../../utils/logger";
 
 @injectable()
 export class BookingController {
@@ -27,6 +27,7 @@ export class BookingController {
     const reqBody = createBookingSchema.parse(req.body);
 
     if (!reqBody) {
+      logger.error("Invalid request body");
       throw createError(400);
     }
 
@@ -38,7 +39,8 @@ export class BookingController {
       booking.createdBy = forUserId;
     } else if (requestingUser?.role === "standard" && forUserId) {
       // standard user cannot create booking for other users
-      throw createError(403, "You cannot create booking for other users");
+      logger.error("Standard user cannot create booking for other users");
+      throw createError(403, "Unauthorized operation");
     } else {
       // admin and standard user can create booking for themselves
       booking.createdBy = requestingUser?.id;
@@ -66,12 +68,13 @@ export class BookingController {
     return res.status(200).send(booking);
   }
 
-  async onUpdateBooking(req: Request, res: Response, next: NextFunction) {
+  async onUpdateBooking(req: Request, res: Response) {
     const requestingUser = req.user;
     const bookingId = stringSchema.parse(req.params.bookingId);
     const booking = updateBookingSchema.parse(req.body);
 
     if (!booking || !bookingId) {
+      logger.error("Invalid request body or bookingId");
       throw createError(400);
     }
 
@@ -80,24 +83,37 @@ export class BookingController {
       createdBy = undefined;
     }
 
-    try {
-      const updatedBooking = await this.bookingInteractor.updateBooking(
-        bookingId,
-        booking,
-        createdBy
-      );
+    const updatedBooking = await this.bookingInteractor.updateBooking(
+      bookingId,
+      booking,
+      createdBy
+    );
 
-      return res.status(200).send(updatedBooking);
-    } catch (error) {
-      if (error instanceof PrismaClientKnownRequestError) {
-        return next(createError(404, "Booking not found"));
-      }
-      return next(error);
-    }
+    return res.status(200).send(updatedBooking);
   }
 
   async onDeleteBooking(req: Request, res: Response) {
     const bookingId = stringSchema.parse(req.params.bookingId);
+    const requestingUser = req.user;
+
+    if (!bookingId) {
+      logger.error("bookingId not found in request params");
+      throw createError(400);
+    }
+
+    const existingBooking = await this.bookingInteractor.getBooking(bookingId);
+    if (!existingBooking) {
+      logger.error("Booking not found");
+      throw createError(404);
+    }
+
+    if (
+      existingBooking.createdBy !== requestingUser?.id &&
+      requestingUser?.role !== "admin"
+    ) {
+      logger.error("User not authorized to delete booking of other user");
+      throw createError(403, "Unauthorized operation");
+    }
 
     await this.bookingInteractor.deleteBooking(bookingId);
